+++ oak-core/src/main/java/org/apache/jackrabbit/oak/core/MutableRoot.java
@@ -42,8 +42,8 @@ import org.apache.jackrabbit.oak.api.CommitFailedException;
 import org.apache.jackrabbit.oak.api.ContentSession;
 import org.apache.jackrabbit.oak.api.QueryEngine;
 import org.apache.jackrabbit.oak.api.Root;
-import org.apache.jackrabbit.oak.api.URLWritableBlob;
-import org.apache.jackrabbit.oak.api.URLWritableBlobRoot;
+import org.apache.jackrabbit.oak.api.blob.URLWritableBlob;
+import org.apache.jackrabbit.oak.api.blob.URLWritableBlobRoot;
 import org.apache.jackrabbit.oak.commons.PathUtils;
 import org.apache.jackrabbit.oak.plugins.index.diffindex.UUIDDiffIndexProviderWrapper;
 import org.apache.jackrabbit.oak.query.ExecutionContext;
