+++ oak-store-spi/src/main/java/org/apache/jackrabbit/oak/plugins/value/jcr/ValueFactoryImpl.java
@@ -39,8 +39,8 @@ import org.apache.jackrabbit.oak.api.Blob;
 import org.apache.jackrabbit.oak.api.PropertyState;
 import org.apache.jackrabbit.oak.api.PropertyValue;
 import org.apache.jackrabbit.oak.api.Root;
-import org.apache.jackrabbit.oak.api.URLWritableBlob;
-import org.apache.jackrabbit.oak.api.URLWritableBlobRoot;
+import org.apache.jackrabbit.oak.api.blob.URLWritableBlob;
+import org.apache.jackrabbit.oak.api.blob.URLWritableBlobRoot;
 import org.apache.jackrabbit.oak.api.binary.URLWritableBinary;
 import org.apache.jackrabbit.oak.api.binary.URLWritableBinaryValueFactory;
 import org.apache.jackrabbit.oak.commons.UUIDUtils;
